refactor: BaseTimeEntity directory 이동 적용
