+++ backend/core/src/main/java/com/rollthedice/backend/domain/debate/entity/DebateRoom.java
@@ -1,7 +1,7 @@
 package com.rollthedice.backend.domain.debate.entity;
 
 import com.rollthedice.backend.domain.member.entity.Member;
-import com.rollthedice.backend.global.config.BaseTimeEntity;
+import com.rollthedice.backend.global.common.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
