+++ backend/core/src/main/java/com/rollthedice/backend/domain/member/entity/Member.java
@@ -1,7 +1,7 @@
 package com.rollthedice.backend.domain.member.entity;
 
 import com.rollthedice.backend.domain.member.dto.MemberServiceDto;
-import com.rollthedice.backend.global.config.BaseTimeEntity;
+import com.rollthedice.backend.global.common.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.*;
