+++ backend/core/src/main/java/com/rollthedice/backend/domain/readNews/entity/ReadNews.java
@@ -2,7 +2,7 @@ package com.rollthedice.backend.domain.readNews.entity;
 
 import com.rollthedice.backend.domain.member.entity.Member;
 import com.rollthedice.backend.domain.news.entity.News;
-import com.rollthedice.backend.global.config.BaseTimeEntity;
+import com.rollthedice.backend.global.common.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
